Resolved some warnings and you can now outline the currently hovered entity.
Former-commit-id: 945ca7e85e2d5eaaa458d994775359bf7e111f93

+++ Application/Game.cpp
@@ -128,7 +128,7 @@ void Game::SaveCameraPosition()
 
 	if (camera_save.is_open())
 	{
-		for (size_t i = 0; i < 16; i++)
+		for (s32 i = 0; i < 16; i++)
 		{
 			camera_save << m_Camera->GetOrientation()[i];
 			camera_save << "\n";
